requirements.txt
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
Pillow==11.3.0
sqlparse==0.5.5
asgiref==3.8.1
